pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "automcp"
version = "0.1.0"
description = "Model Context Protocol server implementation"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.110.0",
    "httpx>=0.27.0",
    "mcp>=1.1.0",
    "pydantic>=2.0.0",
    "pyyaml>=6.0.2",
    "streamlit>=1.41.1",
    "typer>=0.15.1",
    "uvicorn>=0.27.0",
    "websockets>=12.0"
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=4.1.0",
    "black>=24.0.0",
    "isort>=5.13.0",
    "mypy>=1.8.0",
    "ruff>=0.2.0",
    "pre-commit>=3.6.0"
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests", "examples/tests"]
python_files = ["test_*.py"]
addopts = "-v --tb=short"

[tool.black]
line-length = 88
target-version = ["py310"]

[tool.isort]
profile = "black"
line_length = 88
multi_line_output = 3

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
check_untyped_defs = true

[tool.ruff]
line-length = 88
target-version = "py310"
